Image transformation function modified
The image transformation function has been modified so that transformed images are saved in a temporary folder and, therefore, do not appear in the image selection menu. Each time a transformation is requested for a new image, the transformed image is overwritten. There are no issues in case of a new request from another user because, even if the previous user is still viewing the image during a new request, the previously requested image remains visible on the web page until it is closed.